item: went crazy handling different item types
I'm hoping to enumerate the equipment lists for the "equip" command,
but it seems like the data types are complex enough that I'll have to
explicitly handle them. In the meantime, I've taught the item command
how to detect and handle all the various types of items for reasonably
sane display.